requirements.txt
Flask==3.0.3
opencv-python-headless==4.10.0.84
numpy==2.1.1
Pillow==10.4.0
waitress==3.0.0
